:sparkles: feature : 챌린지 참여 관련 API 패키지 위치 변경
Relate to: #4
