+++ src/main/java/com/example/groutine/domain/participation/controller/ChallengeVerificationController.java
@@ -1,0 +1,4 @@
+package com.example.groutine.domain.participation.controller;
+
+public class ChallengeVerificationController {
+}
